Refactor welcome view
